refactor workers
