+++ tests/push/test_segment_worker.py
@@ -4,7 +4,7 @@ import queue
 import pytest
 
 from splitio.api import APIException
-from splitio.push.segmentworker import SegmentWorker, SegmentWorkerAsync
+from splitio.push.workers import SegmentWorker, SegmentWorkerAsync
 from splitio.models.notification import SegmentChangeNotification
 from splitio.optional.loaders import asyncio
